# Conflicts:
#	nemo/collections/asr/modules/rnnt.py
#	nemo/collections/asr/modules/rnnt_abstract.py
#	nemo/collections/asr/parts/submodules/rnnt_loop_labels_computer.py
#	nemo/collections/asr/parts/submodules/tdt_loop_labels_computer.py
#	nemo/collections/asr/parts/utils/rnnt_utils.py

+++ nemo/lightning/io/artifact/file.py
@@ -1,13 +1,28 @@
+# Copyright (c) 2024, NVIDIA CORPORATION.  All rights reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
 import os
 import shutil
 from pathlib import Path
 from typing import Union
+import fiddle as fdl
 
 from nemo.lightning.io.artifact.base import Artifact
 
 
 class PathArtifact(Artifact[Path]):
-    def dump(self, value: Path, absolute_dir: Path, relative_dir: Path) -> Path:
+    def dump(self, instance, value: Path, absolute_dir: Path, relative_dir: Path) -> Path:
         new_value = copy_file(value, absolute_dir, relative_dir)
         return new_value
 
@@ -16,7 +31,10 @@ class PathArtifact(Artifact[Path]):
 
 
 class FileArtifact(Artifact[str]):
-    def dump(self, value: str, absolute_dir: Path, relative_dir: Path) -> str:
+    def dump(self, instance, value: str, absolute_dir: Path, relative_dir: Path) -> str:
+        if not pathize(value).exists():
+            # This is Artifact is just a string.
+            return fdl.Config(FileArtifact, attr=value, skip=True)
         new_value = copy_file(value, absolute_dir, relative_dir)
         return str(new_value)
 
@@ -40,7 +58,7 @@ def copy_file(src: Union[Path, str], path: Union[Path, str], relative_dst: Union
 
 
 class DirArtifact(Artifact[str]):
-    def dump(self, value: str, absolute_dir: Path, relative_dir: Path) -> str:
+    def dump(self, instance, value: str, absolute_dir: Path, relative_dir: Path) -> str:
         value = pathize(value)
         absolute_dir = pathize(absolute_dir)
         relative_dir = pathize(relative_dir)
@@ -58,12 +76,11 @@ class DirArtifact(Artifact[str]):
 
 
 class DirOrStringArtifact(DirArtifact):
-    def dump(self, value: str, absolute_dir: Path, relative_dir: Path) -> str:
+    def dump(self, instance, value: str, absolute_dir: Path, relative_dir: Path) -> str:
         if not pathize(value).exists():
             # This is Artifact is just a string.
-            self.skip = True
-            return value
-        return super().dump(value, absolute_dir, relative_dir)
+            return fdl.Config(DirOrStringArtifact, attr=value, skip=True)
+        return super().dump(instance, value, absolute_dir, relative_dir)
 
     def load(self, path: str) -> str:
         return path
